requirements/base.txt
Django==5.2.12
djangorestframework==3.16.0
psycopg[binary]==3.2.6
gunicorn==23.0.0
whitenoise==6.9.0
